+++ component/frontend/Model/Subscribe.php
@@ -20,6 +20,8 @@ use FOF30\Model\DataModel\Exception\NoItemsFound;
 use FOF30\Model\DataModel\Exception\RecordNotLoaded;
 use FOF30\Model\Model;
 use FOF30\Utils\Ip;
+use Joomla\CMS\Application\ApplicationHelper;
+use Joomla\CMS\Component\ComponentHelper;
 use Joomla\CMS\Environment\Browser as JBrowser;
 use Joomla\CMS\Factory;
 use Joomla\CMS\Filesystem\File;
@@ -29,6 +31,7 @@ use Joomla\CMS\Log\Log as JLog;
 use Joomla\CMS\Plugin\PluginHelper;
 use Joomla\CMS\User\User;
 use Joomla\CMS\User\User as JUser;
+use Joomla\CMS\User\UserHelper;
 use JUserHelper;
 use RuntimeException;
 
@@ -220,19 +223,19 @@ class Subscribe extends Model
 	 *
 	 * @param   bool    $allowNewUser  When true, we can create a new user. False, only update an existing user's data.
 	 * @param   Levels  $level         The subscription level object
+	 * @param   User    $user          The user object to update
 	 *
-	 * @return  bool  True on success
+	 * @return  User|null Updated user object. Null on failure.
 	 */
-	public function updateUserInfo($allowNewUser = true, $level = null)
+	public function updateUserInfo($allowNewUser, ?Levels $level, ?User $user): ?User
 	{
 		$state = $this->getStateVariables();
-		$user = $this->container->platform->getUser();
-		$user = $this->getState('user', $user);
+		$user  = $user ?? $this->container->platform->getUser();
 
 		if (($user->id == 0) && !$allowNewUser)
 		{
 			// New user creation is not allowed. Sorry.
-			return false;
+			throw new RuntimeException('updateUserInfo: New user creation is not allowed');
 		}
 
 		if ($user->id == 0)
@@ -293,7 +296,7 @@ class Subscribe extends Model
 
 					// Remove $user2 and set $user to $user1 so that it gets updated
 					$jUser2 = $this->container->platform->getUser($user2->id);
-					$error = '';
+					$error  = '';
 
 					try
 					{
@@ -310,7 +313,7 @@ class Subscribe extends Model
 						$user2->delete($id2);
 					}
 
-					$user = $this->container->platform->getUser($user1->id);
+					$user        = $this->container->platform->getUser($user1->id);
 					$user->email = $state->email;
 					$user->save(true);
 				}
@@ -332,18 +335,18 @@ class Subscribe extends Model
 		if (is_null($user->id) || ($user->id == 0))
 		{
 			// CREATE A NEW USER
-			$params = array(
+			$params = [
 				'name'      => $state->name,
 				'username'  => $state->username,
 				'email'     => $state->email,
 				'password'  => $state->password,
 				'password2' => $state->password2,
-			);
+			];
 
-			// We have to use JUser directly instead of Factory getUser
-			$user = new JUser(0);
+			// We have to use User directly instead of Factory getUser
+			$user = new User(0);
 
-			$usersConfig = \JComponentHelper::getParams('com_users');
+			$usersConfig = ComponentHelper::getParams('com_users');
 			$newUsertype = $usersConfig->get('new_usertype');
 
 			// get the New User Group from com_users' settings
@@ -352,22 +355,22 @@ class Subscribe extends Model
 				$newUsertype = 2;
 			}
 
-			$params['groups'] = array($newUsertype);
+			$params['groups'] = [$newUsertype];
 
 			$params['sendEmail'] = 0;
 
 			// Set the user's default language to whatever the site's current language is
-			$params['params'] = array(
+			$params['params'] = [
 				'language' => $this->container->platform->getConfig()->get('language'),
-			);
+			];
 
 			// We always block the user, so that only a successful payment or
 			// clicking on the email link activates his account. This is to
 			// prevent spam registrations when the subscription form is abused.
 			$params['block'] = 1;
 
-			$randomString = JUserHelper::genRandomPassword();
-			$hash = \JApplicationHelper::getHash($randomString);
+			$randomString         = UserHelper::genRandomPassword();
+			$hash                 = ApplicationHelper::getHash($randomString);
 			$params['activation'] = $hash;
 
 			$user->bind($params);
@@ -394,18 +397,18 @@ class Subscribe extends Model
 			try
 			{
 				// Delete an existing record
-				$db = $this->container->db;
+				$db    = $this->container->db;
 				$query = $db->getQuery(true)
 					->delete($db->quoteName('#__user_profiles'))
-					->where($db->quoteName('user_id') . ' = ' . (int)($user->id))
+					->where($db->quoteName('user_id') . ' = ' . (int) ($user->id))
 					->where($db->quoteName('profile_key') . ' = ' . $db->q('privacyconsent.privacy'));
 				$db->setQuery($query);
 				$db->execute();
 
 				// Insert a new record
 				$o = (object) [
-					'user_id' => $user->id,
-					'profile_key' => 'privacyconsent.privacy',
+					'user_id'       => $user->id,
+					'profile_key'   => 'privacyconsent.privacy',
 					'profile_value' => 1,
 				];
 				$db->insertObject('#__user_profiles', $o);
@@ -421,16 +424,16 @@ class Subscribe extends Model
 		 */
 		if ($userIsSaved && PluginHelper::isEnabled('system', 'privacyconsent') && class_exists('PlgSystemPrivacyconsent'))
 		{
-			$ip = Ip::getIp();
+			$ip        = Ip::getIp();
 			$userAgent = $this->input->server->get('HTTP_USER_AGENT', '', 'string');
 
 			// Create the user note
-			$userNote = (object) array(
+			$userNote = (object) [
 				'user_id' => $user->id,
 				'subject' => 'PLG_SYSTEM_PRIVACYCONSENT_SUBJECT',
 				'body'    => Text::sprintf('PLG_SYSTEM_PRIVACYCONSENT_BODY', $ip, $userAgent),
 				'created' => Factory::getDate()->toSql(),
-			);
+			];
 
 			try
 			{
@@ -451,7 +454,7 @@ class Subscribe extends Model
 				// Send the activation email
 				if (!isset($params))
 				{
-					$params = array();
+					$params = [];
 				}
 
 				$this->sendActivationEmail($user, $params);
@@ -460,14 +463,18 @@ class Subscribe extends Model
 
 		if (!$userIsSaved)
 		{
-			$this->setState('user', null);
+			$error = 'updateUserInfo: Cannot save user with ID ' . $user->id;
+			$reason = $user->getError();
 
-			return false;
+			if (!empty($reason))
+			{
+				$error = $reason;
+			}
+
+			throw new RuntimeException($error);
 		}
 
-		$this->setState('user', $user);
-
-		return true;
+		return $user;
 	}
 
 	/**
@@ -557,24 +564,33 @@ class Subscribe extends Model
 		// Step #3. Create or update a user record
 		// ----------------------------------------------------------------------
 		$user = $this->container->platform->getUser();
-		$this->setState('user', $user);
-		$userIsSaved = $this->updateUserInfo(true, $level);
 
-		if (!$userIsSaved)
+		try
 		{
-			$this->logSubscriptionCreationFailure(sprintf('Cannot update user information for user ID %d', $user->id));
+			$user = $this->updateUserInfo(true, $level, $user);
 
-			throw new RuntimeException(sprintf('Cannot update user information for user ID %d', $user->id));
+			$this->setState('user', $user);
 		}
+		catch (Exception $e)
+		{
+			$this->logSubscriptionCreationFailure(sprintf('Cannot update user information for user ID %d -- %s', $user->id, $e->getMessage()));
 
-		$user = $this->getState('user', $user);
+			throw new RuntimeException(sprintf($e->getMessage()));
+		}
 
 		// Store the user's ID in the session
 		$this->container->platform->setSessionVar('subscribes.user_id', $user->id, 'com_akeebasubs');
 
 		// Step #3.b. Look for an existing unpaid subscription for the same level and user
 		// ----------------------------------------------------------------------
-		$existingSubscription = $this->findExistingUnpaidSubscription($user->id, $level->getId());
+		try
+		{
+			$existingSubscription = $this->findExistingUnpaidSubscription($user->id, $level->getId());
+		}
+		catch (Exception $e)
+		{
+			$existingSubscription = null;
+		}
 		$useExistingSubscription = false;
 
 		/**
@@ -591,10 +607,41 @@ class Subscribe extends Model
 			// Matching plans: the subscription record is for a recurring subscription and the plan ID matches $recurringId
 			$matchingPlans   = $recurringId == $recurringPlanId;
 
-			// If we have matching characteristics return the old subscription record without update
+			// If we have matching characteristics conditionally return the old subscription record without update
 			if ($bothNull || $matchingPlans)
 			{
-				return $existingSubscription;
+				/**
+				 * Only return the old record if the full price in the record the same as the one from the validation.
+				 *
+				 * This check prevents two cases where the user would be paying a different price than the one presented
+				 * to them.
+				 *
+				 * A. You started a payment which included a time-limited discount, canceled the payment popup and came
+				 * back a few days (less than 7) later when the time-limited discount does not apply. In this case you
+				 * would be told that you need to pay the full price but since we're using the already saved payment ID
+				 * Paddle would ask you to pay the discounted price. This is not right. You lost your chance, you need
+				 * to pay full price.
+				 *
+				 * Important note: we do not need to check explicitly for the 7 days mentioned above. The Subscriptions
+				 * model automatically removes records with state=N (unpaid subs) older than 7 days. Since we go through
+				 * this model in findExistingUnpaidSubscription() we are automatically enforcing this time limit.
+				 *
+				 * B. You forgot to apply a a coupon code. In this case you correctly cancel the payment and retry
+				 * subscribing with the coupon code. You are shown the correct (discounted) price. HOWEVER, because of
+				 * this code above we'd be reusing the previous payment ID which causes Paddle to ask you to pay full
+				 * price again. This is not right, you entered a coupon code, it has to be honored. (This problem did
+				 * actually happen in production).
+				 *
+				 * By checking the price of the validation vs the price of the stored record we can decide if the stored
+				 * payment ID can be reused or not.
+				 */
+				$returnSame = abs($existingSubscription->gross_amount - $validation->price->gross) <= 0.001;
+
+				// Should I return the existing record?
+				if ($returnSame)
+				{
+					return $existingSubscription;
+				}
 			}
 
 			// Otherwise, notify our code we have to update an existing subscription
@@ -1049,7 +1096,7 @@ class Subscribe extends Model
 		}
 
 		// Load the users plugin group.
-		\JPluginHelper::importPlugin('user');
+		PluginHelper::importPlugin('user');
 
 		// Store the data.
 		if (!$user->save())
@@ -1612,12 +1659,12 @@ TEXT;
 	 *
 	 * @since   7.0.0
 	 */
-	public function getRelatedSubscriptions($recurring = true): Collection
+	public function getRelatedSubscriptions($recurring, ?User $user = null): Collection
 	{
 		// Get the user I am interested in
 		/** @var User $user */
-		$user = $this->container->platform->getUser();
-		$user = $this->getState('user', $user);
+		$defaultUser = $this->getState('user', $this->container->platform->getUser());
+		$user = $user ?? $defaultUser;
 
 		// A guest user cannot have subscriptions so let's exit early
 		if ($user->guest)