+++ plugins/system/asexpirationnotify/asexpirationnotify.php
@@ -7,13 +7,20 @@
 
 defined('_JEXEC') or die();
 
-use FOF30\Container\Container;
+use Akeeba\Subscriptions\Admin\Helper\Email;
 use Akeeba\Subscriptions\Admin\Model\Levels;
 use Akeeba\Subscriptions\Admin\Model\Subscriptions;
+use FOF30\Container\Container;
 use FOF30\Date\Date;
+use Joomla\CMS\Component\ComponentHelper;
+use Joomla\CMS\Language\Text;
+use Joomla\CMS\Log\Log;
+use Joomla\CMS\Plugin\CMSPlugin;
 
-class plgSystemAsexpirationnotify extends JPlugin
+class plgSystemAsexpirationnotify extends CMSPlugin
 {
+	protected static $langStringPrefix = 'PLG_SYSTEM_ASEXPIRATIONNOTIFY';
+
 	/**
 	 * Should this plugin be allowed to run? True if FOF can be loaded and the Akeeba Subscriptions component is enabled
 	 *
@@ -26,45 +33,17 @@ class plgSystemAsexpirationnotify extends JPlugin
 	 */
 	public function __construct(& $subject, $config = array())
 	{
+		parent::__construct($subject, $config);
+
 		if (!defined('FOF30_INCLUDED') && !@include_once(JPATH_LIBRARIES . '/fof30/include.php'))
 		{
 			$this->enabled = false;
 		}
 
 		// Do not run if Akeeba Subscriptions is not enabled
-		if (!JComponentHelper::isEnabled('com_akeebasubs'))
+		if (!ComponentHelper::isEnabled('com_akeebasubs'))
 		{
 			$this->enabled = false;
-		}
-
-		if (!is_object($config['params']))
-		{
-			$config['params'] = new JRegistry($config['params']);
-		}
-
-		parent::__construct($subject, $config);
-
-		// Timezone fix; avoids errors printed out by PHP 5.3.3+ (thanks Yannick!)
-		if (function_exists('date_default_timezone_get') && function_exists('date_default_timezone_set'))
-		{
-			if (function_exists('error_reporting'))
-			{
-				$oldLevel = error_reporting(0);
-			}
-
-			$serverTimezone = @date_default_timezone_get();
-
-			if (empty($serverTimezone) || !is_string($serverTimezone))
-			{
-				$serverTimezone = 'UTC';
-			}
-
-			if (function_exists('error_reporting'))
-			{
-				error_reporting($oldLevel);
-			}
-
-			@date_default_timezone_set($serverTimezone);
 		}
 	}
 
@@ -100,13 +79,15 @@ class plgSystemAsexpirationnotify extends JPlugin
 			return;
 		}
 
+		Log::addLogger(['text_file' => "akeebasubs_emails.php"], Log::ALL, ['akeebasubs.emails']);
+
 		$defaultOptions = array(
 			'time_limit' => 2,
 		);
 
 		$options = array_merge($defaultOptions, $options);
 
-		\JLog::add("Starting Expiration Notify - Time limit {$options['time_limit']} seconds", \JLog::DEBUG, "akeebasubs.cron.expirationnotify");
+		Log::add("Starting Expiration Notify - Time limit {$options['time_limit']} seconds", Log::DEBUG, "akeebasubs.cron.expirationnotify");
 
 		// Get today's date
 		$jNow = new Date();
@@ -128,7 +109,7 @@ class plgSystemAsexpirationnotify extends JPlugin
 		/** @var Levels $level */
 		foreach ($levels as $level)
 		{
-			\JLog::add("Processing level " . $level->title, \JLog::DEBUG, "akeebasubs.cron.expirationnotify");
+			Log::add("Processing level " . $level->title, Log::DEBUG, "akeebasubs.cron.expirationnotify");
 
 			// Load the notification thresholds and make sure they are sorted correctly!
 			$notify1     = $level->notify1;
@@ -157,18 +138,18 @@ class plgSystemAsexpirationnotify extends JPlugin
 			$subsModel = Container::getInstance('com_akeebasubs')->factory->model('Subscriptions')->tmpInstance();
 
 			$subs1 = $subsModel->getClone()
-				 ->contact_flag(0)
-				 ->level($level->akeebasubs_level_id)
-				 ->enabled(1)
-				 ->expires_from($jFrom->toSql())
-				 ->expires_to($jTo->toSql())
-				 ->get(true);
+				->contact_flag(0)
+				->level($level->akeebasubs_level_id)
+				->enabled(1)
+				->expires_from($jFrom->toSql())
+				->expires_to($jTo->toSql())
+				->get(true);
 
-			\JLog::add("First Notification - From " . $jFrom->format('Y/m/d H:i:s T') . ' - To ' . $jTo->format('Y/m/d H:i:s T') . ' - Contact flag 0 - Found ' . $subs1->count(), \JLog::DEBUG, "akeebasubs.cron.expirationnotify");
+			Log::add("First Notification - From " . $jFrom->format('Y/m/d H:i:s T') . ' - To ' . $jTo->format('Y/m/d H:i:s T') . ' - Contact flag 0 - Found ' . $subs1->count(), Log::DEBUG, "akeebasubs.cron.expirationnotify");
 
 			// Get the subscriptions expiring within the next $notify2 days for
 			// users which we have contacted only once
-			$subs2 = array();
+			$subs2 = [];
 
 			if ($notify2 > 0)
 			{
@@ -183,11 +164,11 @@ class plgSystemAsexpirationnotify extends JPlugin
 					->expires_to($jTo->toSql())
 					->get(true);
 
-				\JLog::add("Second Notification - From " . $jFrom->format('Y/m/d H:i:s T') . ' - To ' . $jTo->format('Y/m/d H:i:s T') . ' - Contact flag 1 - Found ' . $subs1->count(), \JLog::DEBUG, "akeebasubs.cron.expirationnotify");
+				Log::add("Second Notification - From " . $jFrom->format('Y/m/d H:i:s T') . ' - To ' . $jTo->format('Y/m/d H:i:s T') . ' - Contact flag 1 - Found ' . $subs1->count(), Log::DEBUG, "akeebasubs.cron.expirationnotify");
 			}
 
 			// Get the subscriptions expired $notifyAfter days ago
-			$subs3 = array();
+			$subs3 = [];
 
 			if ($notifyAfter > 0)
 			{
@@ -206,7 +187,7 @@ class plgSystemAsexpirationnotify extends JPlugin
 					->expires_to($jTo->toSql())
 					->get(true);
 
-				\JLog::add("After Expiration Notification - From " . $jFrom->format('Y/m/d H:i:s T') . ' - To ' . $jTo->format('Y/m/d H:i:s T') . ' - Contact flag 3 - Found ' . $subs1->count(), \JLog::DEBUG, "akeebasubs.cron.expirationnotify");
+				Log::add("After Expiration Notification - From " . $jFrom->format('Y/m/d H:i:s T') . ' - To ' . $jTo->format('Y/m/d H:i:s T') . ' - Contact flag 3 - Found ' . $subs1->count(), Log::DEBUG, "akeebasubs.cron.expirationnotify");
 			}
 
 			// If there are no subscriptions, bail out
@@ -219,10 +200,10 @@ class plgSystemAsexpirationnotify extends JPlugin
 				continue;
 			}
 
-			// Check is some of those subscriptions have been renewed. If so, set their contactFlag to 2
-			$realSubs = array();
+			// Check is some of those subscriptions have been renewed. If so, set their contactFlag to 3
+			$realSubs = [];
 
-			foreach (array($subs1, $subs2, $subs3) as $subs)
+			foreach ([$subs1, $subs2, $subs3] as $subs)
 			{
 				/** @var Subscriptions $sub */
 				foreach ($subs as $sub)
@@ -230,7 +211,7 @@ class plgSystemAsexpirationnotify extends JPlugin
 					// Skip the subscription if the contact_flag is already 3
 					if ($sub->contact_flag == 3)
 					{
-						\JLog::add("Skipping #" . $sub->akeebasubs_subscription_id . ', contact flag is 3', \JLog::INFO, "akeebasubs.cron.expirationnotify");
+						Log::add("Skipping #" . $sub->akeebasubs_subscription_id . ', contact flag is 3', Log::INFO, "akeebasubs.cron.expirationnotify");
 
 						continue;
 					}
@@ -254,14 +235,14 @@ class plgSystemAsexpirationnotify extends JPlugin
 
 					if ($renewals->count())
 					{
-						\JLog::add("Skipping #" . $sub->akeebasubs_subscription_id . ', renewals found. Updating contact flag to 3', \JLog::INFO, "akeebasubs.cron.expirationnotify");
+						Log::add("Skipping #" . $sub->akeebasubs_subscription_id . ', renewals found. Updating contact flag to 3', Log::INFO, "akeebasubs.cron.expirationnotify");
 
 						// The user has already renewed. Don't send him an email; just update the row
 						$subsModel->getClone()
-						        ->find($sub->akeebasubs_subscription_id)
-						        ->save([
-							        'contact_flag' => 3
-						        ]);
+							->find($sub->akeebasubs_subscription_id)
+							->save([
+								'contact_flag' => 3,
+							]);
 
 						// Timeout check -- Only if we did make a modification!
 						$clockNow = microtime(true);
@@ -296,46 +277,62 @@ class plgSystemAsexpirationnotify extends JPlugin
 			{
 				$processedCount++;
 
-				// Is it the first or the second contact?
-				if ($sub->enabled && ($sub->contact_flag == 0))
+				/**
+				 * If it's a recurring subscription we are not going to notify the user (they have received emails from
+				 * Paddle) and we're going to set the contact flag to 3.
+				 */
+				if (!empty($sub->cancel_url) && !empty($sub->update_url))
+				{
+					// Recurring subscription -- no contact
+					$data = [
+						'contact_flag'   => 3,
+						'first_contact'  => $mNow,
+						'second_contact' => $mNow,
+						'after_contact'  => $mNow,
+					];
+
+					$result = true;
+				}
+				elseif ($sub->enabled && ($sub->contact_flag == 0))
 				{
 					// First contact
-					$data = array(
+					$data = [
 						'contact_flag'  => 1,
-						'first_contact' => $mNow
-					);
+						'first_contact' => $mNow,
+					];
 
 					$result = $this->sendEmail($sub, 'first');
 				}
 				elseif ($sub->enabled && ($sub->contact_flag == 1))
 				{
 					// Second and final contact
-					$data = array(
+					$data = [
 						'contact_flag'   => 2,
-						'second_contact' => $mNow
-					);
+						'second_contact' => $mNow,
+					];
 
 					$result = $this->sendEmail($sub, 'second');
 				}
 				elseif (!$sub->enabled)
 				{
-					$data = array(
+					// Post-expiration notification
+					$data = [
 						'contact_flag'  => 3,
-						'after_contact' => $mNow
-					);
+						'after_contact' => $mNow,
+					];
 
 					$result = $this->sendEmail($sub, 'after');
 				}
 				else
 				{
-					\JLog::add("I have no idea what to do with subscription #{$sub->akeebasubs_subscription_id}", \JLog::WARNING, "akeebasubs.cron.expirationnotify");
+					Log::add("I have no idea what to do with subscription #{$sub->akeebasubs_subscription_id}", Log::WARNING, "akeebasubs.cron.expirationnotify");
 
 					continue;
 				}
 
 				if ($result)
 				{
-					\JLog::add("-- Updating subscription #{$sub->akeebasubs_subscription_id}, contact flag set to {$data['contact_flag']}", \JLog::DEBUG, "akeebasubs.cron.expirationnotify");
+					Log::add("-- Updating subscription #{$sub->akeebasubs_subscription_id}, contact flag set to {$data['contact_flag']}", Log::DEBUG, "akeebasubs.cron.expirationnotify");
 					$table = $subsModel->getClone();
 					$table->find($sub->akeebasubs_subscription_id);
 					$table->setState('_dontNotify', true);
@@ -349,7 +346,7 @@ class plgSystemAsexpirationnotify extends JPlugin
 				if (($options['time_limit'] > 0) && ($elapsed > $options['time_limit']))
 				{
 					$leftOvers = count($realSubs) - $processedCount;
-					\JLog::add("I ran out of time. Number of subscriptions in queue left unprocessed: $leftOvers", \JLog::DEBUG, "akeebasubs.cron.expirationnotify");
+					Log::add("I ran out of time. Number of subscriptions in queue left unprocessed: $leftOvers", Log::DEBUG, "akeebasubs.cron.expirationnotify");
 
 					// Unset last run timestamp and return
 					$this->setLastRunTimestamp(0);
@@ -476,19 +473,84 @@ class plgSystemAsexpirationnotify extends JPlugin
 
 		// Get a preloaded mailer
 		$key    = 'plg_system_' . $this->_name . '_' . $type;
-		$mailer = \Akeeba\Subscriptions\Admin\Helper\Email::getPreloadedMailer($row, $key);
+		$mailer = Email::getPreloadedMailer($row, $key);
 
-		if ($mailer === false)
+		if (is_null($mailer))
 		{
 			return false;
 		}
 
-		\JLog::add("Sending $type notification to #{$row->akeebasubs_subscription_id} @{$user->username} ($user->name <{$user->email}>)", \JLog::INFO, "akeebasubs.cron.expirationnotify");
+		Log::add("Sending $type notification to #{$row->akeebasubs_subscription_id} @{$user->username} ($user->name <{$user->email}>)", Log::INFO, "akeebasubs.cron.expirationnotify");
 
 		$mailer->addRecipient($user->email);
 		$result = $mailer->Send();
 		$mailer = null;
 
+		// Log the email we just sent
+		$this->logEmail($row, $type);
+
 		return $result;
 	}
+
+	protected function logEmail(Subscriptions $row, $type = '')
+	{
+		$container = Container::getInstance('com_akeebasubs');
+
+		/** @var \Akeeba\Subscriptions\Admin\Model\JoomlaUsers $user */
+		$user = $row->juser ?? $container->factory->model('JoomlaUsers')->tmpInstance()->load($row->user_id);
+		/** @var \Akeeba\Subscriptions\Admin\Model\Levels $level */
+		$level = $row->level ?? $container->factory->model('Levels')->tmpInstance()->load($row->akeebasubs_level_id);
+
+		// Is this a recurring or one-time subscription?
+		$isRecurring   = !empty($row->update_url) && !empty($row->cancel_url);
+		$recurringText = $isRecurring ? 'recurring' : 'one-time';
+
+		// Get a human readable payment state
+		$payState        = $row->getFieldValue('state');
+		$payStateToHuman = [
+			'N' => 'New',
+			'P' => 'Pending',
+			'C' => 'Completed',
+			'X' => 'Canceled',
+		];
+		$payStateHuman   = $payStateToHuman[$payState];
+
+		// Add cancellation reason for canceled subscriptions
+		if ($payState == 'X')
+		{
+			$payStateHuman .= sprintf(' (%s)', $row->cancellation_reason);
+
+			if ($row->cancellation_reason == 'past_due')
+			{
+				$recurringText = 'recurring';
+			}
+		}
+
+		// Create the log entry text
+		$logEntry = sprintf(
+			'Expiration %s (%s) to %s <%s> (%s) for %s #%05u %s -- %s %s to %s -- Contact Flag %d',
+			$type,
+			Text::_(sprintf("%s_%s", self::$langStringPrefix, $type)),
+			$user->username,
+			$user->email,
+			$user->name,
+			$payStateHuman,
+			$row->akeebasubs_subscription_id,
+			$level->title,
+			$recurringText,
+			Date::getInstance($row->publish_up)->format('Y-m-d H:i:s T'),
+			Date::getInstance($row->publish_down)->format('Y-m-d H:i:s T'),
+			$row->contact_flag
+		);
+
+		// If there has been a transaction recorded append it to the log entry
+		if ($payState != 'N')
+		{
+			$logEntry .= sprintf(' -- %s payment key %s', ucfirst($row->processor), $row->processor_key);
+		}
+
+		// Write the log entry
+		JLog::add($logEntry, JLog::INFO, 'akeebasubs.emails');
+	}
+
 }